add:添加中轴线布局

+++ ZFFlowLayout/Src/Common/ZFFlowLayoutMacro.h
@@ -14,12 +14,12 @@
  右对齐布局时：右上角第一个item 距离四周的距离
   ***/
 typedef struct FlowLayoutItemEdgeInsets {
-    CGFloat top, left, bottom, right;  // specify amount to inset (positive) for each of the edges. values can be negative to 'outset'
+    float top, left, bottom, right;  // specify amount to inset (positive) for each of the edges. values can be negative to 'outset'
 } FlowLayoutItemEdgeInsets;
 
 /*!** item所属的组section 距离四周的距离 ***/
 typedef struct FlowLayoutSectionEdgeInsets {
-    CGFloat top, left, bottom, right;  // specify amount to inset (positive) for each of the edges. values can be negative to 'outset'
+    float top, left, bottom, right;  // specify amount to inset (positive) for each of the edges. values can be negative to 'outset'
 } FlowLayoutSectionEdgeInsets;
 
 #endif /* ZFFlowLayoutMacro_h */
